up: update string helper
