+++ src/stdlib/src/Helper/StringHelper.php
@@ -2,8 +2,7 @@
 
 namespace Swoft\Stdlib\Helper;
 
-use Exception;
-use RuntimeException;
+use Swoft\Stdlib\Concern\RandomStringTrait;
 use function explode;
 use function is_string;
 use function lcfirst;
@@ -12,7 +11,6 @@ use function mb_strlen;
 use function mb_strtoupper;
 use function preg_match;
 use function preg_replace;
-use function random_bytes;
 use function str_pad;
 use function str_replace;
 use function strlen;
@@ -21,7 +19,6 @@ use function strrpos;
 use function substr;
 use function substr_replace;
 use function trim;
-use function uniqid;
 use const STR_PAD_LEFT;
 use const STR_PAD_RIGHT;
 
@@ -32,6 +29,8 @@ use const STR_PAD_RIGHT;
  */
 class StringHelper
 {
+    use RandomStringTrait;
+
     /**
      * The cache of snake-cased words.
      *
@@ -86,6 +85,13 @@ class StringHelper
         return static::$camelCache[$value] = ($lcfirst ? lcfirst(static::studly($value)) : static::studly($value));
     }
 
+    /**
+     * @param string $string
+     * @param string $delimiter
+     * @param int    $limit
+     *
+     * @return array
+     */
     public static function toArray(string $string, string $delimiter = ',', int $limit = 0): array
     {
         $string = trim($string, "$delimiter ");
@@ -105,6 +111,13 @@ class StringHelper
         return $values;
     }
 
+    /**
+     * @param string $str
+     * @param string $separator
+     * @param int    $limit
+     *
+     * @return array
+     */
     public static function explode(string $str, string $separator = ',', int $limit = 0): array
     {
         return static::toArray($str, $separator, $limit);
@@ -306,85 +319,6 @@ class StringHelper
     public static function parseCallback(string $callback, string $default): array
     {
         return static::contains($callback, '@') ? explode('@', $callback, 2) : [$callback, $default];
-    }
-
-    /**
-     * @param string $prefix
-     * @param bool   $moreEntropy
-     *
-     * @return string
-     */
-    public static function uniqID(string $prefix = '', bool $moreEntropy = false): string
-    {
-        if (false === $moreEntropy) {
-            return uniqid($prefix, false);
-        }
-
-        return str_replace('.', '', uniqid($prefix, true));
-    }
-
-    /**
-     * Generate a more truly "random" alpha-numeric string.
-     *
-     * @param int $length
-     *
-     * @return string
-     * @throws RuntimeException
-     * @throws Exception
-     */
-    public static function random(int $length = 16): string
-    {
-        $string = '';
-
-        while (($len = strlen($string)) < $length) {
-            $size  = $length - $len;
-            $bytes = random_bytes($size);
-
-            $string .= substr(str_replace(['/', '+', '='], '', base64_encode($bytes)), 0, $size);
-        }
-
-        return $string;
-    }
-
-    /**
-     * Generate a more truly "random" bytes.
-     *
-     * @param int $length
-     *
-     * @return string
-     * @throws Exception
-     * @deprecated since version 5.2. Use random_bytes instead.
-     */
-    public static function randomBytes(int $length = 16): string
-    {
-        if (PHP_MAJOR_VERSION >= 7 || defined('RANDOM_COMPAT_READ_BUFFER')) {
-            $bytes = random_bytes($length);
-        } elseif (function_exists('openssl_random_pseudo_bytes')) {
-            $bytes = openssl_random_pseudo_bytes($length, $strong);
-
-            if ($bytes === false || $strong === false) {
-                throw new RuntimeException('Unable to generate random string.');
-            }
-        } else {
-            throw new RuntimeException('OpenSSL extension or paragonie/random_compat is required for PHP 5 users.');
-        }
-
-        return $bytes;
-    }
-
-    /**
-     * Generate a "random" alpha-numeric string.
-     * Should not be considered sufficient for cryptography, etc.
-     *
-     * @param int $length
-     *
-     * @return string
-     */
-    public static function quickRandom(int $length = 16): string
-    {
-        $pool = '0123456789abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ';
-
-        return substr(str_shuffle(str_repeat($pool, $length)), 0, $length);
     }
 
     /**
@@ -606,112 +540,6 @@ class StringHelper
         return $matches[0];
     }
 
-
-    /**
-     * Generates a random string of a given type and length. Possible
-     * values for the first argument ($type) are:
-     *  - alnum    - alpha-numeric characters (including capitals)
-     *  - alpha    - alphabetical characters (including capitals)
-     *  - hexdec   - hexadecimal characters, 0-9 plus a-f
-     *  - numeric  - digit characters, 0-9
-     *  - nozero   - digit characters, 1-9
-     *  - distinct - clearly distinct alpha-numeric characters.
-     * For values that do not match any of the above, the characters passed
-     * in will be used.
-     * ##### Example
-     *     echo Str::random('alpha', 20);
-     *     // Output:
-     *     DdyQFCddSKeTkfjCewPa
-     *     echo Str::random('distinct', 20);
-     *     // Output:
-     *     XCDDVXV7FUSYAVXFFKSL
-     *
-     * @param string  $type   A type of pool, or a string of characters to use as the pool
-     * @param integer $length Length of string to return
-     *
-     * @return  string
-     * @throws Exception
-     */
-    public static function randomString($type = 'alnum', $length = 8): string
-    {
-        $utf8 = false;
-
-        switch ($type) {
-            case 'alnum':
-                $pool = '0123456789abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ';
-                break;
-            case 'alpha':
-                $pool = 'abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ';
-                break;
-            case 'lowalnum':
-                $pool = '0123456789abcdefghijklmnopqrstuvwxyz';
-                break;
-            case 'hexdec':
-                $pool = '0123456789abcdef';
-                break;
-            case 'numeric':
-                $pool = '0123456789';
-                break;
-            case 'nozero':
-                $pool = '123456789';
-                break;
-            case 'distinct':
-                $pool = '2345679ACDEFHJKLMNPRSTUVWXYZ';
-                break;
-            default:
-                $pool = (string)$type;
-                $utf8 = !self::isAscii($pool);
-                break;
-        }
-
-        // Split the pool into an array of characters
-        $pool = ($utf8 === true) ? self::strSplit($pool, 1) : str_split($pool, 1);
-
-        // Largest pool key
-        $max = count($pool) - 1;
-
-        $str = '';
-        for ($i = 0; $i < $length; $i++) {
-            // Select a random character from the pool and add it to the string
-            $str .= $pool[random_int(0, $max)];
-        }
-
-        // Make sure alnum strings contain at least one letter and one digit
-        if ($type === 'alnum' && $length > 1) {
-            if (ctype_alpha($str)) {
-                // Add a random digit
-                $str[random_int(0, $length - 1)] = chr(random_int(48, 57));
-            } elseif (ctype_digit($str)) {
-                // Add a random letter
-                $str[random_int(0, $length - 1)] = chr(random_int(65, 90));
-            }
-        }
-
-        return $str;
-    }
-
-    /**
-     * Create a simple random token-string
-     *
-     * @param integer $length Length of string
-     * @param string  $salt
-     *
-     * @return  string  Generated token
-     * @throws Exception
-     */
-    public static function randomToken(int $length = 24, string $salt = ''): string
-    {
-        $string = '';
-        $chars  = '0456789abc1def2ghi3jkl';
-        $maxVal = strlen($chars) - 1;
-
-        for ($i = 0; $i < $length; ++$i) {
-            $string .= $chars[random_int(0, $maxVal)];
-        }
-
-        return md5($string . $salt);
-    }
-
     /**
      * @param string $str
      *
@@ -772,25 +600,6 @@ class StringHelper
         }
 
         return $path;
-    }
-
-    /**
-     * @param string $prefix
-     * @param bool   $moreEntropy
-     *
-     * @return string
-     */
-    public static function getUniqid(string $prefix = '', bool $moreEntropy = false): string
-    {
-        // If on Cygwin, $moreEntropy must be TRUE.
-        if (EnvHelper::isCygwin()) {
-            $moreEntropy = true;
-        }
-
-        $uniqId = uniqid($prefix, $moreEntropy);
-        $uniqId = str_replace('.', '', $uniqId);
-
-        return $uniqId;
     }
 
     /**
