+++ src/stdlib/src/Helper/SystemHelper.php
@@ -38,7 +38,7 @@ class SystemHelper extends EnvHelper
      */
     public static function setProcessTitle(string $title): bool
     {
-        if (EnvHelper::isMac() || self::isCygwin()) {
+        if (self::isMac() || self::isCygwin()) {
             return false;
         }
